DAY6 REFACTOR : App는 root파일...
